【Fix PIR Unittest No.399】
