+++ test/deprecated/legacy_test/test_functional_conv3d_transpose_deprecated.py
@@ -22,10 +22,8 @@ import paddle.base.dygraph as dg
 import paddle.nn.functional as F
 from paddle import base
 
-paddle.enable_static()
 
-
-class TestFunctionalConv3DTranspose(TestCase):
+class TestFunctionalConv3DTransposeError(TestCase):
     batch_size = 4
     spatial_shape = (8, 8, 8)
     dtype = "float32"
@@ -35,7 +33,7 @@ class TestFunctionalConv3DTranspose(TestCase):
         self.in_channels = 3
         self.out_channels = 5
         self.filter_shape = 3
-        self.padding = 0
+        self.padding = "not_valid"
         self.stride = 1
         self.dilation = 1
         self.groups = 1
@@ -43,80 +41,28 @@ class TestFunctionalConv3DTranspose(TestCase):
         self.act = "sigmoid"
         self.data_format = "NDHWC"
 
+    def test_exception(self):
+        self.prepare()
+        with self.assertRaises(ValueError):
+            self.static_graph_case()
+
     def prepare(self):
         if isinstance(self.filter_shape, int):
             filter_shape = (self.filter_shape,) * 3
         else:
             filter_shape = tuple(self.filter_shape)
+        self.weight_shape = (
+            self.in_channels,
+            self.out_channels // self.groups,
+        ) + filter_shape
+        self.bias_shape = (self.out_channels,)
 
-        self.weight = np.random.uniform(
-            -1,
-            1,
-            (self.in_channels, self.out_channels // self.groups) + filter_shape,
-        ).astype(self.dtype)
-        if not self.no_bias:
-            self.bias = np.random.uniform(-1, 1, (self.out_channels,)).astype(
-                self.dtype
-            )
-
-        self.channel_last = self.data_format == "NDHWC"
-        if self.channel_last:
-            self.input_shape = (
-                (self.batch_size,) + self.spatial_shape + (self.in_channels,)
-            )
-        else:
-            self.input_shape = (
-                self.batch_size,
-                self.in_channels,
-            ) + self.spatial_shape
-
-        self.input = np.random.uniform(-1, 1, self.input_shape).astype(
-            self.dtype
-        )
-
-    def static_graph_case_1(self):
+    def static_graph_case(self):
         main = base.Program()
         start = base.Program()
         with base.unique_name.guard():
             with base.program_guard(main, start):
-                if self.channel_last:
-                    x = paddle.static.data(
-                        "input",
-                        (-1, -1, -1, -1, self.in_channels),
-                        dtype=self.dtype,
-                    )
-                else:
-                    x = paddle.static.data(
-                        "input",
-                        (-1, self.in_channels, -1, -1, -1),
-                        dtype=self.dtype,
-                    )
-                y = paddle.static.nn.conv3d_transpose(
-                    x,
-                    self.out_channels,
-                    output_size=self.output_size,
-                    filter_size=self.filter_shape,
-                    stride=self.stride,
-                    padding=self.padding,
-                    dilation=self.dilation,
-                    groups=self.groups,
-                    param_attr=paddle.nn.initializer.Assign(self.weight),
-                    bias_attr=False
-                    if self.no_bias
-                    else paddle.nn.initializer.Assign(self.bias),
-                    act=self.act,
-                    data_format=self.data_format,
-                )
-        exe = base.Executor(self.place)
-        exe.run(start)
-        (out,) = exe.run(main, feed={"input": self.input}, fetch_list=[y])
-        return out
-
-    def static_graph_case_2(self):
-        main = base.Program()
-        start = base.Program()
-        with base.unique_name.guard():
-            with base.program_guard(main, start):
+                self.channel_last = self.data_format == "NDHWC"
                 if self.channel_last:
                     x = x = paddle.static.data(
                         "input",
@@ -130,11 +76,11 @@ class TestFunctionalConv3DTranspose(TestCase):
                         dtype=self.dtype,
                     )
                 weight = paddle.static.data(
-                    "weight", self.weight.shape, dtype=self.dtype
+                    "weight", self.weight_shape, dtype=self.dtype
                 )
                 if not self.no_bias:
                     bias = paddle.static.data(
-                        "bias", self.bias.shape, dtype=self.dtype
+                        "bias", self.bias_shape, dtype=self.dtype
                     )
                 y = F.conv3d_transpose(
                     x,
@@ -149,58 +95,61 @@ class TestFunctionalConv3DTranspose(TestCase):
                 )
                 if self.act == 'sigmoid':
                     y = F.sigmoid(y)
-        exe = base.Executor(self.place)
-        exe.run(start)
-        feed_dict = {"input": self.input, "weight": self.weight}
-        if not self.no_bias:
-            feed_dict["bias"] = self.bias
-        (out,) = exe.run(main, feed=feed_dict, fetch_list=[y])
-        return out
-
-    def dygraph_case(self):
-        with dg.guard(self.place):
-            x = paddle.to_tensor(self.input)
-            weight = paddle.to_tensor(self.weight)
-            bias = None if self.no_bias else paddle.to_tensor(self.bias)
-            y = F.conv3d_transpose(
-                x,
-                weight,
-                bias,
-                output_size=self.output_size,
-                padding=self.padding,
-                stride=self.stride,
-                dilation=self.dilation,
-                groups=self.groups,
-                data_format=self.data_format,
-            )
-            if self.act == 'sigmoid':
-                y = F.sigmoid(y)
-            out = y.numpy()
-        return out
-
-    def _test_identity(self):
-        self.prepare()
-        out1 = self.static_graph_case_1()
-        out2 = self.static_graph_case_2()
-        out3 = self.dygraph_case()
-        np.testing.assert_array_almost_equal(out1, out2)
-        np.testing.assert_array_almost_equal(out2, out3)
-
-    def test_identity_cpu(self):
-        self.place = base.CPUPlace()
-        self._test_identity()
-
-    @unittest.skipIf(
-        not base.core.is_compiled_with_cuda(), "core is not compiled with CUDA"
-    )
-    def test_identity_gpu(self):
-        self.place = base.CUDAPlace(0)
-        self._test_identity()
 
 
-class TestFunctionalConv3DTransposeCase2(TestFunctionalConv3DTranspose):
+class TestFunctionalConv3DTransposeErrorCase2(
+    TestFunctionalConv3DTransposeError
+):
     def setUp(self):
         self.in_channels = 3
+        self.out_channels = 5
+        self.filter_shape = 3
+        self.padding = [1, 2, 2, 1, 3]
+        self.stride = 1
+        self.dilation = 1
+        self.groups = 1
+        self.no_bias = False
+        self.act = "sigmoid"
+        self.data_format = "NDHWC"
+
+
+class TestFunctionalConv3DTransposeErrorCase3(
+    TestFunctionalConv3DTransposeError
+):
+    def setUp(self):
+        self.in_channels = 3
+        self.out_channels = 5
+        self.filter_shape = 3
+        self.padding = [[0, 0], [0, 0], [1, 1], [1, 2], [2, 1]]
+        self.stride = 1
+        self.dilation = 1
+        self.groups = 1
+        self.no_bias = False
+        self.act = "sigmoid"
+        self.data_format = "NDHWC"
+
+
+class TestFunctionalConv3DTransposeErrorCase4(
+    TestFunctionalConv3DTransposeError
+):
+    def setUp(self):
+        self.in_channels = 3
+        self.out_channels = 5
+        self.filter_shape = 3
+        self.padding = [[0, 0], [1, 2], [1, 1], [0, 0], [2, 1]]
+        self.stride = 1
+        self.dilation = 1
+        self.groups = 1
+        self.no_bias = False
+        self.act = "sigmoid"
+        self.data_format = "NCDHW"
+
+
+class TestFunctionalConv3DTransposeErrorCase5(
+    TestFunctionalConv3DTransposeError
+):
+    def setUp(self):
+        self.in_channels = -2
         self.out_channels = 5
         self.filter_shape = 3
         self.padding = 0
@@ -212,70 +161,16 @@ class TestFunctionalConv3DTransposeCase2(TestFunctionalConv3DTranspose):
         self.data_format = "NCDHW"
 
 
-class TestFunctionalConv3DTransposeCase3(TestFunctionalConv3DTranspose):
+class TestFunctionalConv3DTransposeErrorCase7(
+    TestFunctionalConv3DTransposeError
+):
     def setUp(self):
         self.in_channels = 4
-        self.out_channels = 6
+        self.out_channels = 5
         self.filter_shape = 3
         self.padding = 0
+        self.output_size = "not_valid"
         self.stride = 1
-        self.dilation = 1
-        self.groups = 2
-        self.no_bias = False
-        self.act = "sigmoid"
-        self.data_format = "NDHWC"
-
-
-class TestFunctionalConv3DTransposeCase4(TestFunctionalConv3DTranspose):
-    def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 6
-        self.filter_shape = 3
-        self.padding = "same"
-        self.stride = 1
-        self.dilation = 1
-        self.groups = 2
-        self.no_bias = True
-        self.act = "sigmoid"
-        self.data_format = "NDHWC"
-
-
-class TestFunctionalConv3DTransposeCase5(TestFunctionalConv3DTranspose):
-    def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 6
-        self.filter_shape = 3
-        self.padding = "valid"
-        self.stride = (1, 2, 1)
-        self.dilation = (2, 1, 1)
-        self.groups = 2
-        self.no_bias = False
-        self.act = "sigmoid"
-        self.data_format = "NDHWC"
-
-
-class TestFunctionalConv3DTransposeCase6(TestFunctionalConv3DTranspose):
-    def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 4
-        self.filter_shape = 3
-        self.padding = "valid"
-        self.stride = (1, 2, 1)
-        self.dilation = 1
-        self.groups = 4
-        self.no_bias = False
-        self.act = "sigmoid"
-        self.data_format = "NDHWC"
-
-
-class TestFunctionalConv3DTransposeCase7(TestFunctionalConv3DTranspose):
-    def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 4
-        self.filter_shape = 3
-        self.padding = "valid"
-        self.output_size = (10, 17, 10)
-        self.stride = (1, 2, 1)
         self.dilation = 1
         self.groups = 1
         self.no_bias = False
@@ -283,54 +178,30 @@ class TestFunctionalConv3DTransposeCase7(TestFunctionalConv3DTranspose):
         self.data_format = "NCDHW"
 
 
-class TestFunctionalConv3DTransposeCase8(TestFunctionalConv3DTranspose):
+class TestFunctionalConv3DTransposeErrorCase8(
+    TestFunctionalConv3DTransposeError
+):
     def setUp(self):
         self.in_channels = 4
-        self.out_channels = 6
+        self.out_channels = 5
         self.filter_shape = 3
-        self.padding = [[0, 0], [1, 2], [1, 2], [2, 1], [0, 0]]
+        self.padding = 0
         self.stride = 1
         self.dilation = 1
-        self.groups = 2
+        self.groups = 1
         self.no_bias = False
         self.act = "sigmoid"
-        self.data_format = "NDHWC"
+        self.data_format = "not_valid"
 
 
-class TestFunctionalConv3DTransposeCase9(TestFunctionalConv3DTranspose):
+class TestFunctionalConv3DTransposeErrorCase9(
+    TestFunctionalConv3DTransposeError
+):
     def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 6
+        self.in_channels = 3
+        self.out_channels = 4
         self.filter_shape = 3
-        self.padding = [[0, 0], [0, 0], [1, 1], [1, 1], [2, 2]]
-        self.stride = 1
-        self.dilation = 1
-        self.groups = 2
-        self.no_bias = False
-        self.act = "sigmoid"
-        self.data_format = "NCDHW"
-
-
-class TestFunctionalConv3DTransposeCase10(TestFunctionalConv3DTranspose):
-    def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 6
-        self.filter_shape = 3
-        self.padding = [1, 1, 2, 2, 1, 1]
-        self.stride = 1
-        self.dilation = 1
-        self.groups = 2
-        self.no_bias = False
-        self.act = "sigmoid"
-        self.data_format = "NCDHW"
-
-
-class TestFunctionalConv3DTransposeCase11(TestFunctionalConv3DTranspose):
-    def setUp(self):
-        self.in_channels = 4
-        self.out_channels = 6
-        self.filter_shape = 3
-        self.padding = [1, 2, 1]
+        self.padding = 0
         self.stride = 1
         self.dilation = 1
         self.groups = 2
@@ -352,37 +223,29 @@ class TestFunctionalConv3DTransposeErrorCase10(TestCase):
         self.groups = 1
         self.data_format = "NCDHW"
 
-    def static_graph_case(self):
-        main = base.Program()
-        start = base.Program()
-        with base.unique_name.guard():
-            with base.program_guard(main, start):
-                x = paddle.static.data(
-                    "input", self.input.shape, dtype=paddle.float32
-                )
-                y = paddle.static.nn.conv3d_transpose(
-                    x,
-                    self.num_filters,
-                    self.filter_size,
-                    stride=self.stride,
-                    padding=self.padding,
-                    dilation=self.dilation,
-                    groups=self.groups,
-                    param_attr=paddle.nn.initializer.Assign(self.filter),
-                    bias_attr=False
-                    if self.bias is None
-                    else paddle.nn.initializer.Assign(self.bias),
-                    act=None,
-                    data_format=self.data_format,
-                )
-        exe = base.Executor()
-        exe.run(start)
-        (out,) = exe.run(main, feed={"input": self.input}, fetch_list=[y])
-        return out
+    def dygraph_case(self):
+        with dg.guard():
+            x = paddle.to_tensor(self.input, dtype=paddle.float32)
+            w = paddle.to_tensor(self.filter, dtype=paddle.float32)
+            b = (
+                None
+                if self.bias is None
+                else paddle.to_tensor(self.bias, dtype=paddle.float32)
+            )
+            y = F.conv3d_transpose(
+                x,
+                w,
+                b,
+                padding=self.padding,
+                stride=self.stride,
+                dilation=self.dilation,
+                groups=self.groups,
+                data_format=self.data_format,
+            )
 
-    def test_static_exception(self):
+    def test_dygraph_exception(self):
         with self.assertRaises(ValueError):
-            self.static_graph_case()
+            self.dygraph_case()
 
 
 class TestFunctionalConv3DTransposeErrorCase11(
